+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Default slot linked to `description` prop in `SModalConfirm` component.
+
+### Fixed
+- `prevent-close` prop in `SModalConfirm` component.
 
 ## 2.3.0 - 2025-06-12
 ### Added